backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
openai>=1.0.0
pandas>=2.0.0
